Fix HeatMapLegend so its bars show real values and not relatives (0-100) Fix bars in FRD widget. Made some adjustments to the widget too. Fix color computation and thresholds in GenePlotter and HeatMap

+++ src/viewOpenGL/HeatMapLegendGL.h
@@ -57,8 +57,8 @@ private:
     // limits and boundaries
     qreal m_lower_threshold;
     qreal m_upper_threshold;
-    qreal m_max;
-    qreal m_min;
+    int m_max;
+    int m_min;
 
     // texture color data
     QGLTexture2D m_texture;
